--- src/main/resources/com/diploma/demo/core/landplot/controller/create-land-plot-tab.fxml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<?import javafx.scene.text.Text?>
-
-<ScrollPane fx:controller="com.diploma.demo.view.landplot.LandPlotControllerCreate"
-            hbarPolicy="NEVER" vbarPolicy="ALWAYS"
-            xmlns:fx="http://javafx.com/fxml">
-    <GridPane alignment="center" hgap="10" vgap="10"  prefHeight="550.0">
-
-        <Text text="Welcome"
-              GridPane.columnIndex="0" GridPane.rowIndex="0"
-              GridPane.columnSpan="2"/>
-
-        <Label text="Appartmente:"
-               GridPane.columnIndex="0" GridPane.rowIndex="1"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="1"
-                fx:id="appartmentTextField"/>
-
-        <Label text="city:"
-               GridPane.columnIndex="0" GridPane.rowIndex="2"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="2"
-                fx:id="cityTextField"/>
-
-        <Label text="home_number:"
-               GridPane.columnIndex="0" GridPane.rowIndex="3"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="3"
-                fx:id="homeNumberTextField"/>
-
-        <Label text="region:"
-               GridPane.columnIndex="0" GridPane.rowIndex="4"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="4"
-                fx:id="regionTextField"/>
-
-        <Label text="street:"
-               GridPane.columnIndex="0" GridPane.rowIndex="5"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="5"
-                fx:id="streetTextField"/>
-
-        <Label text="cadastral_number:"
-               GridPane.columnIndex="0" GridPane.rowIndex="6"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="6"
-                fx:id="cadastralNumberTextField"/>
-
-        <Label text="category:"
-               GridPane.columnIndex="0" GridPane.rowIndex="7"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="7"
-                fx:id="categoryTextField"/>
-
-        <Label text="current_marks:"
-               GridPane.columnIndex="0" GridPane.rowIndex="8"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="8"
-                fx:id="currentMarksTextField"/>
-
-        <Label text="intended_use:"
-               GridPane.columnIndex="0" GridPane.rowIndex="9"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="9"
-                fx:id="intetdedUseTextField"/>
-
-        <Label text="land_plot_purpose:"
-               GridPane.columnIndex="0" GridPane.rowIndex="10"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="10"
-                fx:id="LandPlotPurposeTextField"/>
-
-
-        <Label text="notes:"
-               GridPane.columnIndex="0" GridPane.rowIndex="11"
-        />
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="11"
-                fx:id="notesTextField"/>
-
-
-        <Label text="surface:"
-               GridPane.columnIndex="0" GridPane.rowIndex="12"/>
-
-        <TextField
-                GridPane.columnIndex="1" GridPane.rowIndex="12"
-                fx:id="surfaceTextField"/>
-
-        <!--<Label text="Password:"
-               GridPane.columnIndex="0" GridPane.rowIndex="2"/>
-
-        <PasswordField fx:id="passwordField"
-                       GridPane.columnIndex="1" GridPane.rowIndex="2"/>-->
-
-        <Button text="Create" GridPane.columnIndex="0" GridPane.rowIndex="13"
-                fx:id="createButton" onAction="#click"/>
-    </GridPane>
-</ScrollPane>
