+++ src/main/resources/com/diploma/demo/core/landplot/controller/land-plot-page-tab.fxml
@@ -3,19 +3,133 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-
-<VBox xmlns="http://javafx.com/javafx"
-          xmlns:fx="http://javafx.com/fxml"
-          fx:controller="com.diploma.demo.view.CounterController"
-          prefHeight="400.0" prefWidth="600.0">
+<?import javafx.scene.text.Text?>
+<VBox prefHeight="400.0"
+      prefWidth="600.0"
+      xmlns="http://javafx.com/javafx/11.0.1"
+      xmlns:fx="http://javafx.com/fxml/1"
+      fx:controller="com.diploma.demo.view.landplot.LandPlotControllerRead">
    <children>
       <TabPane tabClosingPolicy="UNAVAILABLE">
          <tabs>
             <Tab fx:id="landPlotsTab" text="landPlots">
-              <fx:include source="read-landPlot-tab.fxml"/>
+               <VBox alignment="CENTER" spacing="10">
+                  <Button fx:id="refreshTableButton" onAction="#refresh">Refresh</Button>
+                  <TableView fx:id="tableView" prefHeight="400.0" prefWidth="1200.0">
+                     <!--<columns>
+                        <TableColumn text="ID">
+                        </TableColumn>
+                        <TableColumn text="Appartamentn">
+                        </TableColumn>
+                        <TableColumn text="City">
+                        </TableColumn>
+                     </columns>-->
+                  </TableView>
+               </VBox>
             </Tab>
             <Tab fx:id="createlandPlotsTab" text="create landPlots">
-              <fx:include source="create-land-plot-tab.fxml"/>
+               <ScrollPane>
+                  <GridPane alignment="center" hgap="10" vgap="10"  prefHeight="550.0">
+
+                     <Text text="Welcome"
+                           GridPane.columnIndex="0" GridPane.rowIndex="0"
+                           GridPane.columnSpan="2"/>
+
+                     <Label text="Appartmente:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="1"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="1"
+                             fx:id="appartmentTextField"/>
+
+                     <Label text="city:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="2"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="2"
+                             fx:id="cityTextField"/>
+
+                     <Label text="home_number:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="3"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="3"
+                             fx:id="homeNumberTextField"/>
+
+                     <Label text="region:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="4"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="4"
+                             fx:id="regionTextField"/>
+
+                     <Label text="street:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="5"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="5"
+                             fx:id="streetTextField"/>
+
+                     <Label text="cadastral_number:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="6"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="6"
+                             fx:id="cadastralNumberTextField"/>
+
+                     <Label text="category:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="7"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="7"
+                             fx:id="categoryTextField"/>
+
+                     <Label text="current_marks:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="8"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="8"
+                             fx:id="currentMarksTextField"/>
+
+                     <Label text="intended_use:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="9"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="9"
+                             fx:id="intetdedUseTextField"/>
+
+                     <Label text="land_plot_purpose:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="10"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="10"
+                             fx:id="LandPlotPurposeTextField"/>
+
+
+                     <Label text="notes:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="11"
+                     />
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="11"
+                             fx:id="notesTextField"/>
+
+
+                     <Label text="surface:"
+                            GridPane.columnIndex="0" GridPane.rowIndex="12"/>
+
+                     <TextField
+                             GridPane.columnIndex="1" GridPane.rowIndex="12"
+                             fx:id="surfaceTextField"/>
+
+
+                     <Button text="Create" GridPane.columnIndex="0" GridPane.rowIndex="13"
+                             fx:id="createButton" onAction="#click"/>
+                  </GridPane>
+               </ScrollPane>
+            </Tab>
+            <Tab>
+
             </Tab>
          </tabs>
