--- src/main/resources/com/diploma/demo/core/landplot/controller/read-landPlot-tab.fxml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<VBox xmlns="http://javafx.com/javafx"
-      xmlns:fx="http://javafx.com/fxml"
-      spacing="10" alignment="CENTER"
-      fx:controller="com.diploma.demo.view.landplot.LandPlotControllerRead">
-    <Button fx:id="refreshTableButton" onAction="#refresh">Refresh</Button>
-    <TableView fx:id="tableView" prefHeight="400.0" prefWidth="1200.0"/>
-</VBox>
