+++ update.sql
@@ -96,3 +96,33 @@ alter table if exists owner add column apartmentn int8
 alter table if exists owner_aud add column apartmentn int8
 alter table if exists owner add column apartmentn int8
 alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
